// packages/source-sync/src/lambda-modules/sync-boundary/infrastructure/adapter/out/db/typeorm/fields/mt-work-order-columns.enum.ts
export enum MtWorkOrderColumns {
  ID = 'ID',
  DOMAIN_ID = 'DOMAIN_ID',
  GRID_EVENT_ID = 'GRID_EVENT_ID',
  PARENT_WORK_ORDER_ID = 'PARENT_WORK_ORDER_ID',
  ROOT_WORK_ORDER_ID = 'ROOT_WORK_ORDER_ID',
  SAP_NOTIFICATION_NUMBER = 'SAP_NOTIFICATION_NUMBER',
  SAP_ORDER_NUMBER = 'SAP_ORDER_NUMBER',
  STAGE = 'STAGE',
  NORMALIZED_STATUS = 'NORMALIZED_STATUS',
  SAP_ESTADO = 'SAP_ESTADO',
  SAP_STATUS_CODE = 'SAP_STATUS_CODE',
  SAP_LAST_UPDATED_AT = 'SAP_LAST_UPDATED_AT',
  SAP_CURSOR_AT = 'SAP_CURSOR_AT',
  DESCRIPTION = 'DESCRIPTION',
  TECHNICAL_LOCATION = 'TECHNICAL_LOCATION',
  SAP_IS_VISIBLE = 'SAP_IS_VISIBLE',
  SAP_CREATION_DATE = 'SAP_CREATION_DATE',
  MANUAL_OVERRIDES = 'MANUAL_OVERRIDES',
  DELETED_AT = 'DELETED_AT',
  CREATED_AT = 'CREATED_AT',
  UPDATED_AT = 'UPDATED_AT',
}